fix(monitor): split page into server/client components; hide thinking dots during streaming
- Extract MonitorClient.tsx ('use client') with ssr:false dynamic import so
  page.tsx can remain a server component and export metadata
- Guard ThinkingIndicator with !streamingText to prevent overlap when
  thinking state lingers after streaming begins
